+++ server/src/index.ts
@@ -1,6 +1,7 @@
 import express from "express";
 import { errorHandler } from "./middleware/error";
 import authRoutes from "./routes/auth";
+import userRoutes from "./routes/user";
 
 const app = express();
 app.use(express.json());
@@ -9,7 +10,8 @@ app.get("/", (req, res) => {
   res.send("Hello World!");
 });
 
-app.use("/auth", authRoutes);
+app.use("/api/auth", authRoutes);
+app.use("/api/users", userRoutes);
 app.use(errorHandler);
 
 const PORT = process.env.PORT || 4000;
