+++ server/src/prisma/client.ts
@@ -1,0 +1,5 @@
+import { PrismaClient } from "../generated/prisma/client";
+
+const prisma = new PrismaClient();
+
+export default prisma;
